[jit] Delete the ErrorReport default constructor
An error report without a SourceRange context is bad, because it doesn't
tell the user where something happened. Delete the default constructor
to make it harder to create errors like this (you can still use a fake
SourceRange if you absolutely need to).

Also clean up the only case where the default constructor was used.

ghstack-source-id: 51eb653d8b10d6f16dab0df2a8f07740454085f3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32879